refactor(template): 126-002 - evacuate feature code from layer folders into features tree
Moves 10 files that belonged conceptually to a feature slice but physically lived under
web-domain/web-application/web-infrastructure/web-server layer folders, into the features
tree, adopting `…Features.<Slice>[.Layer]` namespaces per the migration manifest:

- web-domain/aggregates/profile/profile.cs -> features/profile/profile-domain.cs
  (TimeWarp.Architecture.Aggregates.Profiles -> Features.Profiles.Domain)
- web-domain/aggregates/profile/profile-id.cs -> features/profile/profile-id-domain.cs
  (same namespace change; [TypedId] generator output verified to follow the new namespace)
- web-infrastructure/persistence/ef-principal-store.cs
  -> features/identity/ef-principal-store-infrastructure.cs
  (TimeWarp.Architecture.Persistence -> Features.Identity.Infrastructure)
- web-server/hubs/chat-hub.cs -> features/chat/chat-hub-server.cs
  (TimeWarp.Architecture.Hubs -> Features.Chat)
- web-server/services/chat-hub-service.cs -> features/chat/chat-hub-service-server.cs
  (TimeWarp.Architecture.Services -> Features.Chat)
- web-application/configuration/web-authn-options.cs
  -> features/identity/web-authn-options-application.cs
- web-application/configuration/web-authn-options-validator.cs
  -> features/identity/web-authn-options-validator-application.cs
- web-application/configuration/agent-token-options.cs
  -> features/identity/agent-token-options-application.cs
- web-application/configuration/agent-token-options-validator.cs
  -> features/identity/agent-token-options-validator-application.cs
  (all four: TimeWarp.Architecture.Configuration -> Features.Identity.Application, the
  namespace already used by the identity handlers and other Application-layer stayers)
- web-server/services/agent-token-authentication-handler.cs
  -> features/identity/agent-token-authentication-scheme-server.cs
  (TimeWarp.Architecture.Services -> Features.Identity; renamed away from a literal
  "-handler-server" filename because "handler" is a registered TWA0015 function token
  reserved for the application layer — this is an ASP.NET Core AuthenticationHandler, not
  a mediator handler, so it takes the documented `<name>-<layer>` escape hatch instead)

Hazard fixups from the manifest: H1 (template.json path), H2 (four using sites retargeted to
Features.Profiles.Domain), H3 (added Features.Identity.Infrastructure global using to
web-infrastructure-tests), H5 (removed web-application's now-dead global using for
TimeWarp.Architecture.Configuration), H7 (HowToAddYourAggregate.md paths), H8 (verified
ProfileId's generated TypedId surface compiles under the new namespace). Also added the
Features.Identity.Application global using to web-server and the web-server-integration-tests
project (needed once WebAuthnOptions/AgentTokenOptions left the Configuration namespace —
discovered during the build checkpoint, not called out by the manifest). Reconciled the
Design regions in web-authn-options.cs and agent-token-options.cs, and updated four stray
doc-comment/exception-message path references elsewhere in the repo that pointed at the old
locations.

dev build: 0 warnings, 0 errors.
