+++ source/container-apps/web/web-server/global-usings.cs
@@ -52,9 +52,11 @@ global using TimeWarp.Architecture.Extensions;
 global using TimeWarp.Foundation.Extensions;
 global using TimeWarp.Architecture.Features.Chat;
 global using TimeWarp.Architecture.Features.Identity;
+global using TimeWarp.Architecture.Features.Identity.Application;
 global using TimeWarp.Architecture.Hubs;
 global using TimeWarp.Identity;
 #if(postgres)
+global using TimeWarp.Architecture.Features.Identity.Infrastructure;
 global using TimeWarp.Architecture.Persistence;
 #endif
 global using TimeWarp.Architecture.Services;
