+++ source/foundation/foundation-infrastructure/persistence/aggregate-db-context.cs
@@ -285,8 +285,8 @@ public abstract class AggregateDbContext : DbContext
         (
           $"'{entry.Entity.GetType().Name}' implements IAggregateRoot but has no mapped 'long {VersionPropertyName}' property. " +
           "Aggregate roots must inherit Entity<TId> (source/foundation/foundation-domain/entities/base/entity.cs) so Version is " +
-          "mapped by convention — see the aggregate pattern exemplar (web-domain/aggregates/profile/profile.cs and " +
-          "aggregates/overview.md)."
+          "mapped by convention — see the aggregate pattern exemplar (web/features/profile/profile-domain.cs and " +
+          "web-domain/aggregates/overview.md)."
         );
       }
