+++ tests/container-apps/web/web-domain-tests/global-usings.cs
@@ -1,5 +1,5 @@
 global using Shouldly;
 
 // Solution usings
-global using TimeWarp.Architecture.Aggregates.Profiles;
+global using TimeWarp.Architecture.Features.Profiles.Domain;
 global using TimeWarp.Fixie;
