+++ tests/container-apps/web/web-infrastructure-tests/global-usings.cs
@@ -2,7 +2,8 @@ global using Shouldly;
 
 // Solution usings
 global using Microsoft.EntityFrameworkCore;
-global using TimeWarp.Architecture.Aggregates.Profiles;
+global using TimeWarp.Architecture.Features.Identity.Infrastructure;
+global using TimeWarp.Architecture.Features.Profiles.Domain;
 global using TimeWarp.Architecture.Features.Profiles.Infrastructure;
 global using TimeWarp.Architecture.Persistence;
 global using TimeWarp.Fixie;
