+++ tests/container-apps/web/web-server-integration-tests/GlobalUsings.cs
@@ -10,6 +10,7 @@ global using TimeWarp.Foundation;
 global using TimeWarp.Architecture.Configuration;
 global using TimeWarp.Foundation.Configuration;
 global using TimeWarp.Architecture.Features.Analytics;
+global using TimeWarp.Architecture.Features.Identity.Application;
 global using TimeWarp.Architecture.Testing;
 global using TimeWarp.Foundation.Types;
 global using TimeWarp.Fixie;
